+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support for [Messenger Profile API](https://developers.facebook.com/docs/messenger-platform/messenger-profile).
 
 ## [1.0.11] - 2017-04-13
 ### Added